Utils - разбиение на модули. Абстрактный класс для View. Presenter для Trip.
