+++ src/utils/render.js
@@ -1,0 +1,47 @@
+import Abstract from "../view/abstract.js";
+
+// Функция: рендерит позицию элемента на странице
+export const RenderPosition = {
+  AFTEREND: `afterend`,
+  BEFOREEND: `beforeend`,
+  AFTERBEGIN: `afterbegin`,
+  BEFOREBEGIN: `beforebegin`
+};
+
+export const renderElement = (container, element, place) => {
+  if (container instanceof Abstract) {
+    container = container.getElement();
+  }
+
+  if (element instanceof Abstract) {
+    element = element.getElement();
+  }
+
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case RenderPosition.BEFOREBEGIN:
+      container.before(element);
+      break;
+    case RenderPosition.AFTEREND:
+      container.after(element);
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(element);
+      break;
+  }
+};
+
+
+// Функция: создает элементы на страницы
+// Принцип работы прост:
+// 1. создаём пустой div-блок
+// 2. берём HTML в виде строки и вкладываем в этот div-блок, превращая в DOM-элемент
+// 3. возвращаем этот DOM-элемент
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`); // 1
+  newElement.innerHTML = template; // 2
+
+  return newElement.firstChild; // 3
+};
