+++ src/view/event-edit.js
@@ -1,125 +1,6 @@
 // Форма создания/редактирования (используется одна форма)
-import {createElement, capitalizeFirstCharacter, getLastWord} from "../utils.js";
-import {getDateTimeShort} from "../utils-date-time.js";
-import {eventOffers} from "../mock/event.js";
-
-const createOffersContains = (offersAvailableArr, offersArr) => {
-  let offersAvailable = ``;
-  for (let i = 0; i < offersAvailableArr.length; i++) {
-    let offer = offersAvailableArr[i];
-    let title = offer.title;
-    let price = offer.price;
-    let nameId = `event-offer-${(getLastWord(title).toLowerCase())}`;
-    let checked = ``;
-    if (offersArr.includes(offer)) {
-      checked = `checked`;
-    }
-
-    let offerAvailable = (
-      `<div class="event__offer-selector">
-          <input class="event__offer-checkbox  visually-hidden" id="${nameId}-1" type="checkbox" name="${nameId}" ${checked}>
-            <label class="event__offer-label" for="${nameId}-1">
-              <span class="event__offer-title">${title}</span>
-              &plus;
-              &euro;&nbsp;<span class="event__offer-price">${price}</span>
-            </label>
-      </div>`
-    );
-    offersAvailable += offerAvailable;
-  }
-  return offersAvailable;
-};
-
-const createOffersBlock = (offersAvailableArr, offersArr) => {
-  let block = ``;
-  if (offersAvailableArr.length) {
-    block = (
-      `<section class="event__section  event__section--offers">
-        <h3 class="event__section-title  event__section-title--offers">Offers</h3>
-          <div class="event__available-offers">
-            ${createOffersContains(offersAvailableArr, offersArr)}
-          </div>
-      </section>`);
-  }
-  return block;
-};
-
-const createPicturesContains = (picturesArr) => {
-  let pictures = ``;
-  for (let i = 0; i < picturesArr.length; i++) {
-    let picture = (
-      `<img class="event__photo" src="${picturesArr[i].src}" alt="${picturesArr[i].description}">`
-    );
-    pictures = pictures + picture;
-  }
-  return pictures;
-};
-
-const createDescriptionContains = (description) => {
-  let block = ``;
-  if (description.length) {
-    block = (
-      `<p class="event__destination-description">${description}</p>`);
-  }
-  return block;
-};
-
-const createPicturesBlock = (picturesArr) => {
-  let block = ``;
-  if (picturesArr.length) {
-    block = (
-      `<div class="event__photos-container">
-        <div class="event__photos-tape">
-          ${createPicturesContains(picturesArr)}
-        </div>
-      </div>`);
-  }
-  return block;
-};
-
-const createDescriptionPicturesBlock = (description, picturesArr) => {
-  let block = ``;
-  if (description.length || picturesArr.length) {
-    block = (
-      `<section class="event__section  event__section--destination">
-
-      <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-        ${createDescriptionContains(description)}
-        ${createPicturesBlock(picturesArr)}
-    </section>`);
-  }
-  return block;
-};
-
-const createEventTypeName = (type) => {
-  const eventTo = [`taxi`, `bus`, `train`, `ship`, `transport`, `drive`, `flight`];
-  let eventInTo = ``;
-
-  if (eventTo.includes(type)) {
-    eventInTo = `${capitalizeFirstCharacter(type)} to`;
-  } else {
-    eventInTo = `${capitalizeFirstCharacter(type)} in`;
-  }
-  return eventInTo;
-};
-
-const createEventEditInfo = (point) => {
-
-  const eventEditInfo = {
-    "eventIcon": `img/icons/${point.type}.png`,
-    "eventType": point.type,
-    "eventTypeName": createEventTypeName(point.type),
-    "eventDestinationName": point.destination.name,
-    "eventStartDateTime": getDateTimeShort(point.date_from),
-    "eventEndDateTime": getDateTimeShort(point.date_to),
-    "eventBasePrice": point.base_price,
-    "eventTypeOffersAvailable": eventOffers[point.type],
-    "eventOffersChecked": point.offers,
-    "eventPictures": point.destination.pictures,
-    "eventDescription": point.destination.description,
-  };
-  return eventEditInfo;
-};
+import Abstract from "./abstract.js";
+import {createEventEditInfo, createOffersBlock, createDescriptionPicturesBlock} from "../utils/event-edit.js";
 
 const createEventEditTemplate = (point) => {
   const {eventIcon, eventTypeName, eventDestinationName, eventStartDateTime, eventEndDateTime, eventBasePrice, eventTypeOffersAvailable, eventOffersChecked, eventPictures, eventDescription} = createEventEditInfo(point);
@@ -244,25 +125,25 @@ const createEventEditTemplate = (point) => {
   );
 };
 
-export default class EventEdit {
+export default class EventEdit extends Abstract {
   constructor(point) {
+    super();
     this._point = point;
-    this._element = null;
+
+    this._formSubmitHandler = this._formSubmitHandler.bind(this);
   }
 
   getTemplate() {
     return createEventEditTemplate(this._point);
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
+  _formSubmitHandler(evt) {
+    evt.preventDefault();
+    this._callback.formSubmit();
   }
 
-  removeElement() {
-    this._element = null;
+  setFormSubmitHandler(callback) {
+    this._callback.formSubmit = callback;
+    this.getElement().querySelector(`form`).addEventListener(`submit`, this._formSubmitHandler);
   }
 }
