+++ src/view/filter.js
@@ -1,5 +1,5 @@
 // Фильтры
-import {createElement} from "../utils.js";
+import Abstract from "./abstract.js";
 
 const createFilterTemplate = () => {
   return (
@@ -24,24 +24,8 @@ const createFilterTemplate = () => {
   );
 };
 
-export default class Filter {
-  constructor() {
-    this._element = null;
-  }
-
+export default class Filter extends Abstract {
   getTemplate() {
     return createFilterTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
