+++ src/view/main-menu.js
@@ -1,5 +1,5 @@
 // Меню
-import {createElement} from "../utils.js";
+import Abstract from "./abstract.js";
 
 const createMainMenuTemplate = () => {
   return (
@@ -10,24 +10,8 @@ const createMainMenuTemplate = () => {
   );
 };
 
-export default class MainMenu {
-  constructor() {
-    this._element = null;
-  }
-
+export default class MainMenu extends Abstract {
   getTemplate() {
     return createMainMenuTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
