+++ src/view/sort.js
@@ -1,5 +1,5 @@
 // Сортировка
-import {createElement} from "../utils.js";
+import Abstract from "./abstract.js";
 
 const createSortTemplate = () => {
   return (
@@ -30,24 +30,8 @@ const createSortTemplate = () => {
   );
 };
 
-export default class Sort {
-  constructor() {
-    this._element = null;
-  }
-
+export default class Sort extends Abstract {
   getTemplate() {
     return createSortTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
